Bucket 권한 관리 및 에디터 파일 업로드 구현 및 파일업로드

+++ module-common/src/main/java/com/zzikza/springboot/web/domain/FileAttribute.java
@@ -1,20 +1,16 @@
 package com.zzikza.springboot.web.domain;
 
 import com.zzikza.springboot.web.domain.enums.EFileStatus;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
 
 @Setter
 @Getter
 @Embeddable
-public class FileAttribute{
+public class FileAttribute {
 
     @Column(name = "FL_NM")
     public String fileName;
@@ -23,27 +19,18 @@ public class FileAttribute{
     public String fileSourceName;
 
     @Column(name = "FL_SZ")
-    public int fileSize;
+    public Long fileSize;
+
     @Column(name = "FL_EXT")
     public String fileExt;
+
     @Column(name = "FL_PTH")
     public String filePath;
-//    @Column(name = "FL_THB_")
-//    protected String fileThumbPath;
+
     @Column(name = "FL_ORD")
     public Integer fileOrder;
+
     @Column(name = "FL_STAT_CD")
     public EFileStatus fileStatus;
 
-//    @Builder
-//    public FileAttribute(String fileName, String fileSourceName, int fileSize, String fileExt, String filePath, String fileThumbPath, int fileOrder, EFileStatus fileStatus) {
-//        this.fileName = fileName;
-//        this.fileSourceName = fileSourceName;
-//        this.fileSize = fileSize;
-//        this.fileExt = fileExt;
-//        this.filePath = filePath;
-//        this.fileThumbPath = fileThumbPath;
-//        this.fileOrder = fileOrder;
-//        this.fileStatus = fileStatus;
-//    }
 }
